desiredIstio: Enable GIE if InferencePool found
Configure the Istio control-plane to enable Gateway API Inference
Extension (GIE) if the InferencePool CRD from GIE exists on the cluster.

Conditionally set one new environment variable in the Istio CR:

    ENABLE_GATEWAY_API_INFERENCE_EXTENSION

This commit resolves NE-2104.

https://issues.redhat.com/browse/NE-2104

* pkg/operator/controller/gatewayclass/controller.go
(inferencepoolCrdName, inferencepoolExperimentalCrdName): New consts.
(NewUnmanaged): Watch customresourcedefinitions, with a predicate on the
InferencePool CRD, in order to reconcile the Istio CR when the
InferencePool CRD is created or deleted.
* pkg/operator/controller/gatewayclass/istio.go (ensureIstio): Use the
new inferencepoolCrdExists helper to check whether the InferencePool CRD
exists, and pass a flag to desiredIstio when the CRD is present.
(inferencepoolCrdExists): New helper.  Using the new crdExists helper,
check whether the InferencePool CRD exists in the
inference.networking.k8s.io or inference.networking.x-k8s.io API groups.
(crdExists): New helper.
(desiredIstio): Add a Boolean parameter, and set the environment
variable when the argument is true.

[do-not-merge]Add OSSM channel and version override annotations

Define annotations for the gatewayclass by which the channel and version
for the OSSM subscription can be overridden.

When creating the subscription, the operator now checks the following
annotations:

    unsupported.do-not-use.openshift.io/ossm-channel
    unsupported.do-not-use.openshift.io/ossm-version

One or both annotations may be specified to override the channel or
version from the config.

This commit implements NE-2108.

https://issues.redhat.com/browse/NE-2108

* pkg/operator/controller/gatewayclass/controller.go
(subscriptionVersionOverrideAnnotationKey)
(subscriptionChannelOverrideAnnotationKey): New consts.
(Reconcile): Check the annotations on the gatewayclass.  Pass the
channel and version to ensureServiceMeshOperatorSubscription.
* pkg/operator/controller/gatewayclass/subscription.go
(ensureServiceMeshOperatorSubscription): Add parameters for the
subscription channel and version, and use the arguments instead of using
the config.