Add OSSM and Istio version override annotations
Define annotations for the gatewayclass by which the catalog ource,
channel, and version for the OSSM subscription and the Istio version can
be overridden.

When creating the subscription or Istio CR, the operator now checks the
following annotations:

    unsupported.do-not-use.openshift.io/ossm-catalog
    unsupported.do-not-use.openshift.io/ossm-channel
    unsupported.do-not-use.openshift.io/ossm-version
    unsupported.do-not-use.openshift.io/istio-version

One or more annotations may be specified to override the catalog source,
channel, OSSM version, or Istio version from the operator config.

This commit implements NE-2108.

https://issues.redhat.com/browse/NE-2108

* hack/ossm-overrides.md: New file.  Document the new annotations.
* pkg/operator/controller/gatewayclass/controller.go
(subscriptionCatalogOverrideAnnotationKey)
(subscriptionChannelOverrideAnnotationKey)
(subscriptionVersionOverrideAnnotationKey)
(istioOverrideAnnotationKey): New consts.
(Reconcile): Check the annotations on the gatewayclass.  Pass the
OSSM catalog source, channel, and version to
ensureServiceMeshOperatorSubscription and to
r.ensureServiceMeshOperatorInstallPlan.
Pass the Istio version to ensureIstio.
* pkg/operator/controller/gatewayclass/controller_test.go
(Test_Reconcile): Add test cases for overriding the Istio version and
for overriding the OSSM catalog source, channel, and version.
* pkg/operator/controller/gatewayclass/subscription.go
(ensureServiceMeshOperatorSubscription): Add parameters for the
subscription channel and version, and use the arguments instead of using
the config.
(ensureServiceMeshOperatorInstallPlan): Add parameter for the version,
and pass the arguemnt to currentInstallPlan.
(currentInstallPlan): Add parameter for the version, and use the
argument instead of using the config to match the desired installplan.
* pkg/operator/controller/gatewayclass/subscription_test.go
(Test_ensureServiceMeshOperatorInstallPlan): Update call to
ensureServiceMeshOperatorInstallPlan.