NE-2097: e2e - Add TestOSSMOperatorUpgradeViaIntermediateVersions
The `TestOSSMOperatorUpgradeViaIntermediateVersions` test verifies that the OSSM operator
correctly upgrades through intermediate versions. It ensures the upgrade logic automatically
approves each InstallPlan along the upgrade graph until the desired version is reached.

This test is not added to the suite of all tests because it will be run
in a dedicated CI job to avoid conflicts with existing GatewayAPI jobs
which install the latest supported version of OSSM operator and Istio.