OCPBUGS-59139: Increase assertExpectedDNSRecords timeout to 2 minutes
This PR addresses OCPBUGS-59139.

https://issues.redhat.com//browse/OCPBUGS-59139

Increase the timeout value for the assertExpectedDNSRecords
test/e2e/util_gatewayapi_test function to 2 minutes to give
the DNSRecords more time to be cleaned up in the e2e test cases
that check for the record deletion. This is to avoid CI flakes.

Also increase the polling interval from 1s to 5s to avoid
unnecessary polls and reduce the noise in the logs.

Modify the assertExpectedDNSRecords function to not error out
on listing the DNSRecords when there is an external issue
and retry the request instead.