OCPBUGS-60859: Fix logic on gatewayapi test cleanup
The current logic of testGatewayAPIDNSListenerUpdate adds the cleanup
routine to the stack before checking if the creation of the gateway
was successful.

This causes a panic during the assertGatewaySuccessful and a subsequent
panic on the cleanup, was the gateway resource is null because of the previous
error.

This fix moves the cleanup to happen just if there was no error when creating
the gateway resource