OCPBUGS-55649: Remove SetEIPForNLBIngressController feature gate
This change removes the Feature Gate SetEIPForNLBIngressController since
it is enabled on Openshift since version v4.18.0.

Additionally it fixes all of the tests to assume that there is no feature gate anymore,
and removes tests that asserts that this feature gate is disabled.

This feature is now available by default to any users