Delete dnsManagementPolicy upgrade logic
Delete upgrade logic for the dnsManagementPolicy field that was added in
OpenShift 4.12.  This upgrade logic was added with the intention that it
would be deleted in OpenShift 4.13.  The deleted logic ensured that the
status.endpointPublishingStrategy.loadBalancer.dnsManagementPolicy field
was set on an IngressController object even if that IngressController
had been created on an earlier OpenShift release on which the field did
not exist.  All clusters from OpenShift 4.13 onward now have the field
set, either by this upgrade logic or by the admission logic, and so the
upgrade logic is no longer needed.

Follow-up to commit 5c36287713ceef7632d85103c8d2fe7e4ad91f14.

* pkg/operator/controller/ingress/controller.go (Reconcile): Delete
dnsManagementPolicy upgrade logic for OpenShift 4.12.