Add godoc for IngressController reconciliation
Add godoc to the ingress controller's Reconcile method to describe the
reconciliation logic at a high level.  In particular, note that
controller's admission logic initializes the IngressController object's
status, and that ensureIngressController and related methods therefore
may assume that the status on which they depend has been initialized.

* pkg/operator/controller/ingress/controller.go (Reconcile): Describe
the reconciliation logic at a high level.