[NE-2183] Implement Gateway status controller
This commit introduces the new Gateway Status controller. This
controller will be responsible on adding specific Openshift conditions
to a Gateway resource managed by CIO, existing on 'openshift-ingress'
namespace.

The conditions added are similar to the ones from Ingress resources,
allowing admins to detect when a LoadBalancer or a DNSRecord where provisioned,
and in case of failure, what failure happened.