Add PrometheusRule for Gateway API gathering
This PR adds a PromtheusRule to collect the usage of Gateway on
an Openshift cluster, based on the GatewayClass. It reduces the metric to:
* Gateways that are accepted
* Tags by the GatewayClass (openshift/not-openshift)

As the PrometheusRule can be added without any caveat to the cluster
(eg.: even if no Gateway is enabled), it is being added as its own rule,
but as part of CIO installation.

JIRA: https://issues.redhat.com/browse/NE-2203