CORS-4229: Revert "dns/gcp: Allow configuring custom endpoints"
This reverts commit 95b36d0270996954500ee1840203c305b66f5431.

The revert is occurring because the Openshift Installer will be creating a Private Zone where
the traffic to *.googleapis.com will go to the ip address of a private service connect endpoint.
The changes for the Openshift Installer can be found here: https://github.com/openshift/installer/pull/9992.
This means that we do not need to explicitly override the api endpoints in each cluster component.

For example a proxy server log from a bastion host shows this:
1762508378.713  91253 10.0.0.5 TCP_TUNNEL/200 11968 CONNECT iamcredentials.googleapis.com:443 - HIER_DIRECT/10.1.0.100 -
1762508378.970  90258 10.0.0.5 TCP_TUNNEL/200 5988 CONNECT compute.googleapis.com:443 - HIER_DIRECT/10.1.0.100 -

We can see that the ip address for the default endpoints above is now going to 10.1.0.100 which was the
ip address of the private service connect endpoint. This achieves the same result as setting the API
Endpoint override to https://compute-<endpoint-name>.p.googleapis.com/.