gatewayclass: Enable Horizontal Pod Autoscaling
Enable Horizontal Pod Autoscaling (HPA) on Istio.  Hard-code the
autoscaling parameters based on the cluster infrastructure config:

- If the infrastructure topology is "SingleReplica", set minimum to 1.
- Otherwise, set minimum to 2.
- In any case, set maximum to 10.

* pkg/operator/controller/gatewayclass/controller.go
(gatewayclassControllerIndexFieldName): New const.
(NewUnmanaged): Watch infrastructures.  Initialize fieldIndexer in the
reconciler so that it can be used to create an index over gatewayclasses
later.  The index cannot be created directly in NewUnmanaged as the
gatewayclasses resource might not yet exist when NewUnmanaged is called.
(reconciler): Add fieldIndexer and startGatewayclassControllerIndex.
(Reconcile): Get the cluster infrastructure object and pass it to
ensureIstio.  Create an index over gatewayclasses by
spec.controllerName, using fieldIndexer from the reconciler,
gatewayclassControllerIndexFieldName for the index field name, and
startGatewayclassControllerIndex to ensure the index is only created
once, on first reconciliation.
* pkg/operator/controller/gatewayclass/controller_test.go
(Test_Reconcile): Add test cases for missing cluster infrastructure
config.  Add a test case for SingleReplica topology mode.  Add a test
case with multiple gatewayclasses. Add the cluster infrastructure config
object to existingObjects for the existing test cases. Add the expected
HPA configuration to the expected Istio CRs in test cases.  Initialize
fieldIndexer with a fake indexer in the reconciler.
(FakeIndexer, (FakeIndexer).IndexField): New type and method,
used in Test_Reconcile.
* pkg/operator/controller/gatewayclass/istio.go (ensureIstio): Add an
infraConfig parameter, and pass the argument to desiredIstio.  Use the
new index to list gatewayclasses with the OpenShift gateway controller
name, and pass the list of gatewayclasses to desiredIstio as well.
* pkg/operator/controller/gatewayclass/istio.go (desiredIstio): Add an
infraConfig parameter and a gatewayclasses parameter, and use the
arguments to look up the infrastructure topology mode and and configure
HPA accordingly.